requirements_platform.txt
# Requirements for BTC Trading Strategy Platform GUI
# Updated: 2025-11-13

# GUI Framework
PySide6>=6.5.0

# Core Trading
alpaca-trade-api>=3.0.0
alpaca-py>=0.9.0
pandas>=2.0.0
numpy>=1.24.0

# Backtesting Framework
backtesting>=0.6.0
vectorbt>=0.27.0

# Technical Analysis
ta>=0.10.0
ta-lib>=0.4.0

# Optimization
scikit-optimize>=0.9.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Machine Learning & Statistics
hmmlearn>=0.3.0
statsmodels>=0.14.0

# Visualization
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Reporting
reportlab>=4.0.0

# Configuration & Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
tqdm>=4.65.0
python-dateutil>=2.8.0
pytz>=2023.0

# Security & Encryption
cryptography>=41.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: Enhanced Performance
numba==0.58.0

# Build Tools
pyinstaller>=6.0.0